pyproject.toml
[project]
name = "datacom-ai"
version = "0.1.0"
description = "Production-grade AI platform for retrieval-augmented pipelines and reasoning agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov",
    "black",
    "ruff",
]
ui = [
    "gradio>=4.0.0",
]
eval = [
    "deepeval>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

